+++ src/interval.h
@@ -7,19 +7,19 @@
 
 class interval {
   public:
-    double min;
-    double max;
+    float min;
+    float max;
 
     HOST_DEVICE interval() : min(INFINITY), max(-INFINITY) {}
 
-    HOST_DEVICE interval(double mn, double mx) : min(mn), max(mx) {}
+    HOST_DEVICE interval(float mn, float mx) : min(mn), max(mx) {}
 
     /** Accepted ray parameter range [min, max) (half-open on the right). */
-    HOST_DEVICE bool surrounds(double t) const {
+    HOST_DEVICE bool surrounds(float t) const {
       return t >= min && t < max;
     }
 
-    HOST_DEVICE bool contains_closed(double t) const {
+    HOST_DEVICE bool contains_closed(float t) const {
       return min <= t && t <= max;
     }
 };
